Fix] custom interest typo

+++ server/src/main/java/com/ntumart/dipapp/api/repository/ListingRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import com.ntumart.dipapp.models.Interest;
 import com.ntumart.dipapp.models.Product;
 
-public interface ListingRepository extends JpaRepository<Product, Long> {
+public interface ListingRepository extends JpaRepository<Product, Integer> {
 
   @Query(value = "SELECT * FROM PRODUCT WHERE name LIKE CONCAT ('%', :name, '%')", nativeQuery = true)
   public List<Product> getProductsBySearch(@Param("name") String name);
